+++ tests/Instances/Data/Tuple.hs
@@ -1,6 +1,3 @@
-{-# LANGUAGE CPP                #-}
-{-# LANGUAGE DeriveGeneric      #-}
-{-# LANGUAGE StandaloneDeriving #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 {-|
@@ -18,10 +15,6 @@ module Instances.Data.Tuple () where
 import Data.Orphans ()
 import Instances.Utils.GenericArbitrary (genericArbitrary)
 import Test.QuickCheck (Arbitrary(..))
-
-#if !(MIN_VERSION_base(4,16,0))
-import GHC.Generics (Generic)
-#endif
 
 instance ( Arbitrary a
          , Arbitrary b
@@ -102,12 +95,3 @@ instance ( Arbitrary a
          , Arbitrary o
          ) => Arbitrary (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) where
     arbitrary = genericArbitrary
-
--- TODO: Replace these instances with generic-deriving
-#if !(MIN_VERSION_base(4,16,0))
-deriving instance Generic (a, b, c, d, e, f, g, h, i, j, k)
-deriving instance Generic (a, b, c, d, e, f, g, h, i, j, k, l)
-deriving instance Generic (a, b, c, d, e, f, g, h, i, j, k, l, m)
-deriving instance Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n)
-deriving instance Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)
-#endif