+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next [????.??.??]
+* Require `generic-deriving-1.14.1` or later in the test suite.
+
 ### 3.9.1 [2021.08.29]
 * Require `base-orphans-0.8.5` or later in the test suite.